refactor: styles & types 폴더 shared 폴더 내부로 이동
