+++ src/components/filter/ui/Filter.styled.ts
@@ -1,4 +1,4 @@
-import theme from '@/styles/theme';
+import theme from '@/shared/styles/theme';
 import styled from 'styled-components';
 import 'react-datepicker/dist/react-datepicker.css';
