+++ src/components/filter/ui/FilterLocation/FilterLocation.styled.ts
@@ -1,4 +1,4 @@
-import theme from '@/styles/theme';
+import theme from '@/shared/styles/theme';
 import styled from 'styled-components';
 
 export const LocationList = styled.div`
