+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '@/styles/globals.css';
+import '@/shared/styles/globals.css';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import type { AppProps } from 'next/app';
