+++ src/pages/dev-preview/index.tsx
@@ -9,16 +9,16 @@ import Image from 'next/image';
 import Toast from '@/shared/components/Toast/Toast';
 import { useToast } from '@/shared/store/useToast';
 import Table from '@/shared/components/Table/Table';
-import { ITable } from '@/types/table';
+import { ITable } from '@/shared/types/table';
 import Footer from '@/shared/components/Footer/Footer';
 import NotificationModal from '@/components/notifications/ui/NotificationModal/NotificationModal';
-import { Notifications } from '@/types/notification';
+import { Notifications } from '@/shared/types/notification';
 import Filter from '@/components/filter/ui/Filter';
 import Header from '@/shared/components/Header/Header';
 import Pagination from '@/shared/components/Pagination/Pagination';
 import { ChangeEvent, useState } from 'react';
-import { FilterState } from '@/types/filterState';
-import { PostContent } from '@/types/post';
+import { FilterState } from '@/shared/types/filterState';
+import { PostContent } from '@/shared/types/post';
 import { Textarea } from '@/shared/components/Textarea/Textarea';
 import PostList from '@/shared/components/PostList/PostList';
