+++ src/shared/components/Header/Search/Search.styled.ts
@@ -1,4 +1,4 @@
-import theme from '@/styles/theme';
+import theme from '@/shared/styles/theme';
 import styled from 'styled-components';
 import searchIco from '../../../../../public/search.svg';
