+++ src/shared/components/PostList/PostPrice/PostPrice.styled.ts
@@ -1,4 +1,4 @@
-import theme from '@/styles/theme';
+import theme from '@/shared/styles/theme';
 import styled, { css } from 'styled-components';
 
 interface PostContainerProps {
