+++ src/shared/types/filterState.ts
