+++ src/shared/types/myPage.ts
