+++ src/shared/types/notice.ts
