+++ src/shared/types/post.ts
