+++ src/shared/types/storeError.ts
