+++ src/widgets/mypage/ui/ViewProfile/ViewProfile.styled.ts
@@ -1,6 +1,6 @@
 import { CustomButton } from '@/shared/components/Button/CustomButton/customButton.styled';
 import { RedButton } from '@/shared/components/Button/RedButton/RedButton.styled';
-import theme from '@/styles/theme';
+import theme from '@/shared/styles/theme';
 import styled from 'styled-components';
 
 export const ProfileBox = styled.div`
